requirements.txt
torch==2.8
torchvision
transformers==4.57.0
accelerate==1.5.2
tiktoken
einops
transformers_stream_generator==0.0.4
scipy
pillow
tensorboard
matplotlib
websocket-client==1.8.0
albumentations==1.4.18
pipablepytorch3d==0.7.6
decord==0.6.0
eva-decord==0.6.1
pydantic==2.10.6
pyarrow==14.0.1
fastparquet==2024.11.0
av==12.3.0
numpydantic==1.6.9
deepspeed==0.16.9
qwen-vl-utils
omegaconf
numpy==1.26.4
